+++ package.json
@@ -6,8 +6,7 @@
     "shared",
     "packages/*",
     "scripts",
-    "cypress",
-    "firestore-security-rule-tests"
+    "cypress"
   ],
   "private": true,
   "main": "lib/index.js",
