+++ packages/security-rules/tests/howtos.spec.ts
@@ -2,6 +2,7 @@ require('dotenv').config()
 import * as testing from '@firebase/rules-unit-testing';
 import { readFileSync } from 'fs';
 import { setLogLevel, doc, getDoc, setDoc, deleteDoc } from 'firebase/firestore'
+import { pathToFirestoreRules } from './constants';
 
 const { initializeTestEnvironment, assertFails, assertSucceeds } = testing;
 
@@ -17,7 +18,7 @@ describe(DOCUMENT_BASE, () => {
 
         testEnv = await initializeTestEnvironment({
             projectId: process.env.PROJECT_ID,
-            firestore: { rules: readFileSync('../firestore.rules', 'utf8') },
+            firestore: { rules: readFileSync(pathToFirestoreRules, 'utf8') },
         });
     })
